texture_cache: Address ameerj's review
